requirements.txt

fastapi
uvicorn
openai
anthropic
python-dotenv
httpx
requests
pydantic
jinja2
aiofiles

